fix: canonicalize StartPostgres schema bootstrap
